Add support for apr < 1.5.
apr_sockaddr_is_wildcard() was introduced in apr 1.5, but some systems
are still running earlier versions. So for those we provide an
implementation ourselves, which is a direct copy-paste of the
implementation in apr >= 1.5.

This closes issue #7 and obsoletes pull request #4.

Thanks to mamanguy and earsdown for initial patch suggestions.